requirements.txt
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=1.0.0
pandas>=1.5.0
pillow>=9.0.0
timm>=0.9.0
datasets>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0
tensorboard>=2.0.0 
scikit-learn